refactor(commands): Layer 4 — group-owned command areas under commands/groups/ (core, session, config, debug, project, skills, memory, utility); each group owns its handler files and a dispatch slice chained by execute(); parsing, registry, user-command precedence and fallback stay in mod.rs. Refs #2791. Refs #2870.
