+++ crates/tui/src/commands/groups/config/status.rs
