+++ crates/tui/src/commands/groups/core/anchor.rs
