+++ crates/tui/src/commands/groups/core/hooks.rs
