+++ crates/tui/src/commands/groups/core/provider.rs
