+++ crates/tui/src/commands/groups/core/queue.rs
