+++ crates/tui/src/commands/groups/core/stash.rs
