+++ crates/tui/src/commands/groups/debug/balance.rs
