+++ crates/tui/src/commands/groups/debug/debug.rs
