+++ crates/tui/src/commands/groups/memory/memory.rs
