+++ crates/tui/src/commands/groups/memory/note.rs
