+++ crates/tui/src/commands/groups/project/init.rs
