+++ crates/tui/src/commands/groups/session/rename.rs
