+++ crates/tui/src/commands/groups/skills/restore.rs
