+++ crates/tui/src/commands/groups/utility/mcp.rs
