+++ crates/tui/src/commands/groups/utility/network.rs
